src/backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
requests
pydantic